code optimisation and further code development

+++ routes/message.js
@@ -1,69 +1,162 @@
 const express = require("express");
-const { body, validationResult } = require("express-validator");
+const { param, body, validationResult } = require("express-validator");
 var multer = require('multer');
 const path = require('node:path');
 const mongoose = require('mongoose');
 const check_conversation = require("../middleware/conversation");
-const fetch_user = require("../middleware/fetchuser");
+const fetchUser = require("../middleware/fetchUser");
 const conversationModel = require("../models/conversation");
 const userModel = require("../models/user");
 const messageModel = require("../models/messages");
+const groupModel = require("../models/group");
 
 // creating router
 const router = express.Router();
 
 // middleware
-router.use(fetch_user);
+router.use(fetchUser);
+
+// function to send personal message 
+const sendPersonalMessage = async (req, res) =>{
+	
+	if (!req.body.conversation_id) {
+        // Checking conversation from 1st point of view
+        let conversation1 = await conversationModel.findOne({
+            conversation_by: req.body.id,
+            conversation_with: req.body.message_for
+		});
+
+        // Checking conversation from 2nd point of view
+        let conversation2 = await conversationModel.findOne({
+			conversation_by: req.body.message_for,
+			conversation_with: req.body.id
+        });
+
+        // If both points of view not exist
+        if (!conversation1 && !conversation2) {
+			let conversation = await conversationModel({
+				conversation_by: req.body.id,
+				conversation_with: req.body.message_for,
+			});
+
+            // Saving conversation
+            await conversation.save();
+
+            // Attaching conversation id 
+            req.body.conversation_id = conversation._id;
+		} else {
+			// If 1st or 2nd point of view exists
+			req.body.conversation_id = conversation1 ? conversation1._id : conversation2._id;
+        }
+	}
+
+    // Conversation id found in request
+    else if (req.body.conversation_id) {
+		// Finding the conversation with the provided id
+		let conversation = await conversationModel.findById(req.body.conversation_id);
+		
+		// If conversation with that id not exists
+		if (!conversation) {
+			res.status(400).json({ message: "Invalid conversation id" });
+			return false;
+		}
+    }
+	
+	const message_for = await userModel.findById(req.body.message_for);
+			
+	// if message for user not found
+	if(!message_for){
+		return res.status(400).json({ message: "Enter valid message for id" });
+	}
+
+	// conversation info to be updated
+	const update = {
+		last_conversation_at: Date.now(),
+		last_conversation_message: req.body.message
+	};
+
+	// updating the conversation
+	const updateConversation = await conversationModel.findByIdAndUpdate(req.body.conversation_id, update, { new: true });
+			
+	// creating the message model
+	const message = await messageModel({
+		conversation_id: req.body.conversation_id,
+		conversation_type: "Conversation",
+		message_by: req.body.id,
+		message_for: message_for,
+		message: req.body.message,
+		message_status: "sent",
+	});
+	
+	// saving the message
+	await message.save();
+	res.status(200).json({ message: "sent" });
+}
+		
+
+// function to send group message
+const sendGroupMessage = async (req, res) => {
+	if(!req.body.conversation_id){
+		return res.status(400).json({ message: "Enter group id" });
+	}
+	
+	const groupId = req.body.conversation_id;
+	
+	const group = await groupModel.findById(groupId);
+	
+	// if group not found
+	if(!group){
+		return res.status(400).json({ message: "Group not found"});
+	}
+	
+	// creating the message model
+	const message = await messageModel({
+		conversation_id: req.body.conversation_id,
+		conversation_type: "Group",
+		message_by: req.body.id,
+		message_for: group.group_memebers,
+		message: req.body.message,
+		message_status: "sent",
+	});
+	
+	// saving message
+	await message.save();
+	res.status(200).json({ message: "sent" });
+}
 
 // router to send message
-router.post("/sendMessage", check_conversation,
+router.post("/sendMessage",
+	body("conversation_type", "Enter value of Conversation type").isLength({ min: 1 }),
     body("id").custom((value) => {
         if (!mongoose.Types.ObjectId.isValid(value)) {
             return Promise.reject("Enter a valid user id");
         }
         return Promise.resolve();
     }),
-    body("conversation_with").custom((value) => {
-        if (!mongoose.Types.ObjectId.isValid(value)) {
-            return Promise.reject("Enter a valid conversation with id");
-        }
-        return Promise.resolve();
-    }),
     body("message", "Enter a message").isLength({ min: 1 }),
-    async (req, res) => {
+    async (req, res, next) => {
 
         // assigining the validation result
         const errors = validationResult(req);
 
         // error are not empty
         if (!errors.isEmpty()) {
-            return res.status(400).json({ errors: errors.array() })
+            return res.status(400).json({ errors: errors.array() });
         }
 
         try {
-            // creating a filter
-            const filter = { _id: req.body.conversation_id };
-
-            // data info
-            const update = {
-                last_conversation_at: Date.now(),
-                last_conversation_message: req.body.message
-            };
-
-            // updating the data
-            const updateConversation = await conversationModel.findByIdAndUpdate(filter, update, { new: true });
-
-            // saving the message
-            const message = await messageModel({
-                conversation_id: req.body.conversation_id,
-                messaage_by: req.body.id,
-                messaage_for: req.body.conversation_with,
-                message: req.body.message,
-                message_status: "sent",
-            });
-
-            await message.save();
-            return res.status(200).send("message sent");
+			// personal message
+			if(req.body.conversation_type === "Conversation" || req.body.conversation_type === "conversation"){
+				return sendPersonalMessage(req, res);	
+			}
+			// group message
+			else if(req.body.conversation_type === "Group" || req.body.conversation_type === "group"){
+				return sendGroupMessage(req, res);
+			}
+			// invalid conversation type
+			else{
+				return res.json({ message: "Invalid conversation type" });
+			}
         }
         catch (error) {
             console.error(error);
@@ -73,8 +166,8 @@ router.post("/sendMessage", check_conversation,
     });
 
 // router to get multiple conversation list
-router.get("/getConversations",
-    body("id").custom((value) => {
+router.get("/getConversations/:id",
+    param("id").custom((value) => {
         if (!mongoose.Types.ObjectId.isValid(value)) {
             return Promise.reject("Enter a valid user id");
         }
@@ -91,36 +184,40 @@ router.get("/getConversations",
         }
 
         try {
+			
+			const userId = req.params.id;
+			
+			// empty conversation data list
+            const conversationData = [];
+			
             // getting conversation
             let conversations = await conversationModel.find({
                 $or: [
-                    { conversation_by: req.body.id },
-                    { conversation_with: req.body.id }
+                    { conversation_by: userId },
+                    { conversation_with: userId }
                 ]
             }).select("-created_at -__v").exec();
 
-            // empty conversation data list
-            const conversationData = [];
+            if(conversations){
+				// looping the conversations one by one
+				for (const conversation of conversations) {
+					let otherUserId = conversation.conversation_by;
+					if (otherUserId == req.params.id) {
+						otherUserId = conversation.conversation_with;
+					}
 
-            // looping the conversations one by one
-            for (const conversation of conversations) {
-                let otherUserId = conversation.conversation_by;
-                if (otherUserId === req.body.id) {
-                    otherUserId = conversation.conversation_with;
-                }
+					const user = await userModel.findById(otherUserId);
 
-                const user = await userModel.findById(otherUserId);
-
-                // if user exists
-                if (user) {
-                    // adding add to the array
-                    conversationData.push({
-                        conversation_with_name: user.name,
-                        conversation: conversation
-                    });
-                }
-            }
-
+					// if user exists
+					if (user) {
+						// adding add to the array
+						conversationData.push({
+							conversation_with_name: user.name,
+							conversation: conversation
+						});
+					}
+				}
+			}
             return res.status(200).json(conversationData);
         }
         catch (error) {
@@ -130,14 +227,8 @@ router.get("/getConversations",
     });
 
 // router to get previous messages which belong to particular conversation
-router.get("/getPreviousMessages",
-    body("id").custom((value) => {
-        if (!mongoose.Types.ObjectId.isValid(value)) {
-            return Promise.reject("Enter a valid user id");
-        }
-        return Promise.resolve();
-    }),
-    body("conversation_id").custom((value) => {
+router.get("/getPreviousMessages/:conversation_id",
+    param("conversation_id").custom((value) => {
         if (!mongoose.Types.ObjectId.isValid(value)) {
             return Promise.reject("Enter a valid conversation id");
         }
@@ -155,9 +246,11 @@ router.get("/getPreviousMessages",
 
         try {
             // getting message
-            let messages = await messageModel.find({ conversation_id: req.body.conversation_id }).exec();
-
-            return res.status(200).json(messages);
+            let messages = await messageModel.find({ conversation_id: req.params.conversation_id }).exec();
+			if(messages){
+				return res.status(200).json(messages);
+			}
+			return res.status(200).json([]);
         }
         catch (error) {
             console.error(error);
@@ -166,7 +259,7 @@ router.get("/getPreviousMessages",
     });
 
 // router to get new messages from every conversation user have
-router.get("/getNewMessage", body("id").custom((value) => {
+router.get("/getNewMessage/:id", param("id").custom((value) => {
     if (!mongoose.Types.ObjectId.isValid(value)) {
         return Promise.reject("Enter a valid user id");
     }
@@ -186,19 +279,21 @@ router.get("/getNewMessage", body("id").custom((value) => {
 
             // getting all new messages
             let newmessages = await messageModel.find({
-                messaage_for: req.body.id,
+                message_for: req.params.id,
                 $or: [
                     { message_status: "sent" },
                     { message_status: "delivered" }
                 ]
             }).exec();
 
-            // looping each message one by one to update the message status from sent to delivered
-            for (const newmessage of newmessages) {
-                const filter = { _id: newmessage._id };
-                const update = { message_status: "delivered" };
-                let changemessagestatus = await messageModel.findByIdAndUpdate(filter, update, { new: true });
-            }
+			if(newmessages){
+				// looping each message one by one to update the message status from sent to delivered
+				for (const newmessage of newmessages) {
+					const filter = { _id: newmessage._id };
+					const update = { message_status: "delivered" };
+					let changemessagestatus = await messageModel.findByIdAndUpdate(filter, update, { new: true });
+				}
+			}
 
             // returing new messages
             return res.status(200).json({ newmessages });
@@ -228,17 +323,26 @@ router.put("/changeMessageStatus",
         }
 
         try {
-
             // filtering message by id
             const filter = { _id: req.body.message_id };
 
-            // update content
-            const update = { message_status: "read" };
+			// getting message
+			const message = await messageModel.findById(filter);
+			
+			// if id match with message for id then update the message
+			if(req.body.id == message.message_for){
+				// update content
+				const update = { message_status: "read" };
 
-            // updating the message
-            let updatestatus = await messageModel.findByIdAndUpdate(filter, update, { new: this.true });
+				// updating the message
+				let updatestatus = await messageModel.findByIdAndUpdate(filter, update, { new: this.true });
 
-            return res.status(200).json({ message: "Updated status" });
+				return res.status(200).json({ message: "Updated status" });
+			}
+			// if not matched then don't update the message
+			else{
+				return res.status(401).json({ message: "Problem while updating the status of message. Check you have provided correct id which is equal to message_for of that message" });
+			}
         }
         catch (error) {
             console.error(error);
@@ -246,16 +350,17 @@ router.put("/changeMessageStatus",
         }
     });
 
+
 // router to upload a file
-var abc = "";
+var pathString = "";
 var storage = multer.diskStorage({
     destination: function (req, file, cb) {
         cb(null, './docs')
     },
     filename: function (req, file, cb) {
         var datetimestamp = Date.now();
-        abc = datetimestamp + '.' + file.originalname.split('.')[file.originalname.split('.').length - 1];
-        cb(null, abc);
+        pathString = datetimestamp + '.' + file.originalname.split('.')[file.originalname.split('.').length - 1];
+        cb(null, pathString);
     }
 });
 
@@ -266,7 +371,7 @@ var upload = multer({
     }
 }).single('file');
 
-router.post("/sendFile", async (req, res) => {
+router.post("/sendFile", check_conversation, async (req, res) => {
     try {
         await new Promise((resolve, reject) => {
             upload(req, res, function (err) {
@@ -277,18 +382,30 @@ router.post("/sendFile", async (req, res) => {
                 }
             });
         });
-
+		
+		let newPath = path.join('/docs', pathString);
         const message = await messageModel({
             conversation_id: req.body.conversation_id,
-            messaage_by: req.body.id,
-            messaage_for: req.body.conversation_with,
-            message: "/docs/" + abc,
+            conversation_type: "Conversation",
+			message_by: req.body.id,
+            message_for: req.body.message_for,
+            message: newPath,
             message_status: "sent",
             message_type: "file",
         });
 
         await message.save();
-        return res.status(200).json(message._id);
+		
+		const conversationId = { _id: req.body.conversation_id };
+
+        const update = {
+            last_conversation_at: Date.now(),
+            last_conversation_message: newPath
+		};
+		
+        const updateConversation = await conversationModel.findByIdAndUpdate(conversationId, update, { new: true });
+		
+        return res.status(200).end();
     }
     catch (error) {
         console.error(error);
@@ -296,10 +413,9 @@ router.post("/sendFile", async (req, res) => {
     }
 });
 
-
 // router to get a file
-router.get("/getFile",
-    body("message_id").custom((value) => {
+router.get("/getFile/:message_id",
+    param("message_id").custom((value) => {
         if (!mongoose.Types.ObjectId.isValid(value)) {
             return Promise.reject("Enter valid message id");
         }
@@ -318,7 +434,7 @@ router.get("/getFile",
         try {
 
             // getting message id
-            let message_id = req.body.message_id;
+            let message_id = req.params.message_id;
 
             // fetching message
             let message = await messageModel.findById(message_id);
@@ -344,8 +460,8 @@ router.get("/getFile",
     });
 
 // router to download the file
-router.get("/downloadFile",
-    body("message_id").custom((value) => {
+router.get("/downloadFile/:message_id",
+    param("message_id").custom((value) => {
         if (!mongoose.Types.ObjectId.isValid(value)) {
             return Promise.reject("Enter valid message id");
         }
@@ -364,7 +480,7 @@ router.get("/downloadFile",
         try {
 
             // getting message id
-            let message_id = req.body.message_id;
+            let message_id = req.params.message_id;
 
             // fetching message
             let message = await messageModel.findById(message_id);
@@ -399,25 +515,30 @@ router.delete("/deleteConversation",
     }),
     async (req, res) => {
 
-        // assigining the validation result
+        // Assign the validation result
         const errors = validationResult(req);
 
-        // error are not empty
+        // If errors are not empty, return 400 status with error details
         if (!errors.isEmpty()) {
-            return res.status(400).json({ errors: errors.array() })
+            return res.status(400).json({ errors: errors.array() });
         }
 
         try {
-            // destrucing conversation id
-            let conversation_id = req.body.conversation_id;
+            // Destructure conversation id from URL parameters
+            const { conversation_id } = req.body;
 
-            // finding conversation in collection conversations and deleting it
+            // Find and delete the conversation
             const conversation = await conversationModel.findByIdAndDelete(conversation_id);
 
-            // finding messages in collection messages and deleting it
-            const message = await messageModel.deleteMany({ conversation_id: conversation_id });
+            // Check if conversation is not found
+            if (!conversation) {
+                return res.status(404).json({ message: "Conversation not found" });
+            }
 
-            // returning satus code
+            // Delete associated messages
+            const messages = await messageModel.deleteMany({ conversation_id: conversation_id });
+
+            // Return 200 status with success message
             return res.status(200).json({ message: "Conversation and messages deleted" });
         }
         catch (error) {
